+++ Client/Models/ExpensesModel.cs
@@ -3,24 +3,24 @@
 public class ExpensesModel
 {
     public int Id { get; set; }
-    public float Housing { get; set; }
-    public float Groceries { get; set; }
-    public float Utilities { get; set; }
-    public float Vacation { get; set; }
-    public float Transportation { get; set; }
-    public float Medicine { get; set; }
-    public float Clothing { get; set; }
-    public float Media { get; set; }
-    public float Insuranses { get; set; }
-    public float TrackedHousing { get; set; }
-    public float TrackedGroceries { get; set; }
-    public float TrackedUtilities { get; set; }
-    public float TrackedVacation { get; set; }
-    public float TrackedTransportation { get; set; }
-    public float TrackedMedicine { get; set; }
-    public float TrackedClothing { get; set; }
-    public float TrackedMedia { get; set; }
-    public float TrackedInsuranses { get; set; }
+    public decimal Housing { get; set; }
+    public decimal Groceries { get; set; }
+    public decimal Utilities { get; set; }
+    public decimal Vacation { get; set; }
+    public decimal Transportation { get; set; }
+    public decimal Medicine { get; set; }
+    public decimal Clothing { get; set; }
+    public decimal Media { get; set; }
+    public decimal Insuranses { get; set; }
+    public decimal TrackedHousing { get; set; }
+    public decimal TrackedGroceries { get; set; }
+    public decimal TrackedUtilities { get; set; }
+    public decimal TrackedVacation { get; set; }
+    public decimal TrackedTransportation { get; set; }
+    public decimal TrackedMedicine { get; set; }
+    public decimal TrackedClothing { get; set; }
+    public decimal TrackedMedia { get; set; }
+    public decimal TrackedInsuranses { get; set; }
     public DateTime Date { get; set; }
     public int MonthId { get; set; }
     public int YearId { get; set; }
