+++ DAL/Models/DetailedBudgetModel.cs
@@ -6,10 +6,10 @@ namespace DAL.Models
         public DateTime Date { get; set; }
         public string Type { get; set; } = "";
         public string Category { get; set; } = "";
-        public double Amount { get; set; }
+        public decimal Amount { get; set; }
         public string Details { get; set; } = "";
-        public double Balance { get; set; }
-        public double Tracked { get; set; }
+        public decimal Balance { get; set; }
+        public decimal Tracked { get; set; }
         public int MonthId { get; set; }
         public int YearId { get; set; }
     }
